@@ -8,8 +8,8 @@
   <packaging>pom</packaging>
   <version>1.0-SNAPSHOT</version>
   <properties>
-    <spring-boot.version>2.4.2</spring-boot.version>
-    <spring-cloud.version>2020.0.1</spring-cloud.version>
+    <spring-boot.version>2.6.12</spring-boot.version>
+    <spring-cloud.version>2021.0.4</spring-cloud.version>
     <spring-cloud-alibaba.version>2021.1</spring-cloud-alibaba.version>
     <spring-cloud-starter-bootsratp.version>3.1.0</spring-cloud-starter-bootsratp.version>
     <hutool.version>5.7.19</hutool.version>
