@@ -3,8 +3,13 @@ package com.cuukenn.cloud.auth.service.impl;
 import cn.hutool.cache.Cache;
 import cn.hutool.cache.CacheUtil;
 import cn.hutool.core.util.IdUtil;
+import com.cuukenn.cloud.auth.config.CaptchaProperties;
 import com.cuukenn.cloud.auth.service.ICaptchaService;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.stereotype.Service;
+
+import java.time.temporal.ChronoUnit;
 
 /**
  * 内存验证码存储器
@@ -12,8 +17,14 @@ import org.springframework.stereotype.Service;
  * @author changgg
  */
 @Service
+@ConditionalOnMissingBean(ICaptchaService.class)
+@EnableConfigurationProperties(CaptchaProperties.class)
 public class InMemoryCaptchaServiceImpl implements ICaptchaService {
-    private final Cache<String, String> cache = CacheUtil.newFIFOCache(1000, 60 * 1000);
+    private final Cache<String, String> cache;
+
+    public InMemoryCaptchaServiceImpl(CaptchaProperties properties) {
+        cache = CacheUtil.newFIFOCache(1000, properties.getTimeout().get(ChronoUnit.MILLIS));
+    }
 
     @Override
     public String removeIfAbstract(String id) {
