@@ -1,5 +1,6 @@
 package com.cuukenn.openstudysource.cloud.user.converter;
 
+import com.cuukenn.openstudysource.cloud.framework.entity.BaseConverter;
 import com.cuukenn.openstudysource.cloud.user.dto.AuthUserDto;
 import com.cuukenn.openstudysource.cloud.user.dto.UserDto;
 import com.cuukenn.openstudysource.cloud.user.entity.User;
@@ -10,15 +11,7 @@ import org.mapstruct.Mapping;
  * @author changgg
  */
 @Mapper(componentModel = "spring")
-public interface UserConverter {
-    /**
-     * user2dto
-     *
-     * @param user user
-     * @return dto
-     */
-    UserDto toUserDto(User user);
-
+public interface UserConverter extends BaseConverter<UserDto, User> {
     /**
      * user2dto
      *
@@ -28,12 +21,4 @@ public interface UserConverter {
     @Mapping(target = "accountExpired", expression = "java(user.isAccountInvalid())")
     @Mapping(target = "passwordExpired", expression = "java(user.isPasswordInvalid())")
     AuthUserDto toAuthUserDto(User user);
-
-    /**
-     * dto2user
-     *
-     * @param dto dto
-     * @return user
-     */
-    User fromUserDto(UserDto dto);
 }
